Base class for Reference mutations.
