+++ packages/documentator/src/Markdown/Mutations/Reference/Inline.php
@@ -2,15 +2,11 @@
 
 namespace LastDragon_ru\LaraASP\Documentator\Markdown\Mutations\Reference;
 
-use Iterator;
-use LastDragon_ru\LaraASP\Documentator\Markdown\Contracts\Mutation;
 use LastDragon_ru\LaraASP\Documentator\Markdown\Data\Content;
 use LastDragon_ru\LaraASP\Documentator\Markdown\Data\Location;
-use LastDragon_ru\LaraASP\Documentator\Markdown\Data\Reference;
 use LastDragon_ru\LaraASP\Documentator\Markdown\Document;
 use LastDragon_ru\LaraASP\Documentator\Markdown\Extensions\Reference\Node as ReferenceNode;
 use LastDragon_ru\LaraASP\Documentator\Markdown\Utils;
-use League\CommonMark\Extension\CommonMark\Node\Inline\AbstractWebResource;
 use League\CommonMark\Extension\CommonMark\Node\Inline\Image;
 use League\CommonMark\Extension\CommonMark\Node\Inline\Link;
 use Override;
@@ -20,11 +16,7 @@ use function rawurldecode;
 /**
  * Inlines all references.
  */
-class Inline implements Mutation {
-    public function __construct() {
-        // empty
-    }
-
+readonly class Inline extends Mutation {
     /**
      * @inheritDoc
      */
@@ -53,25 +45,6 @@ class Inline implements Mutation {
 
             if ($text !== null) {
                 yield [$location, $text !== '' ? $text : null];
-            }
-        }
-    }
-
-    /**
-     * @return Iterator<array-key, AbstractWebResource|ReferenceNode>
-     */
-    private function nodes(Document $document): Iterator {
-        // Just in case
-        yield from [];
-
-        // Search
-        foreach ($document->node->iterator() as $node) {
-            if ($node instanceof AbstractWebResource && Reference::get($node) !== null) {
-                yield $node;
-            } elseif ($node instanceof ReferenceNode) {
-                yield $node;
-            } else {
-                // empty
             }
         }
     }
